fix: 🐛 fix nbt exporting
when calling nbt[""] we were failing hard (issue #1) so I have changed the code to just export the entire nbt and that seems to work just fine. it isn't clear if this is going to make different JSON because I don't have old JSON to reference. Also added a few more sanity checks for files exiting during parsing.